+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * github-216: [BUG] Floats should be represented as FLOAT and not REAL in MySQL/MariaDB
 * github-217: Support collations for creating/migrating JDBC tables
 * github-218: [BUG] Postgres dialect should be used for Postgres JDBC URLs
+* github-219: [BUG] SchemaMapping should retain incoming comments
 
 ## Breaking changes